OmniParser2/pyproject.toml
[project]
name = "omniparser2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "similarities>=1.2.3",
    "pymilvus==2.5.11",
    "clip==1.0",
    "cos-python-sdk-v5>=1.9.39",
    "einops==0.8.0",
    "fastapi[standard]>=0.124.4",
    "huggingface-hub>=0.23.0",
    "rtree>=1.4.1",
    "loguru>=0.7.3",
    "minio>=7.2.20",
    "peft>=0.18.0",
    "pillow>=12.0.0",
    "timm>=1.0.22",
    "torch>=2.8.0",
    "transformers>=4.57.3",
    "ultralytics>=8.3.240",
    "pydantic-settings>=2.12.0",
    "paddleocr>=3.3.2",
    "paddlepaddle==3.2.2; sys_platform != 'linux'",
    "paddlepaddle-gpu==3.2.2; sys_platform == 'linux'",
    "supervision>=0.27.0",
    "datasets==4.4.1",
    "log-reporter>=1.2.2",
]
[tool.uv]
# 强制覆盖依赖版本，touch版本2.8.0 和 paddlepaddle-gpu版本3.2.2 存在下面依赖冲突
override-dependencies = [
    "nvidia-cublas-cu12==12.8.4.1; sys_platform == 'linux'",
    "nvidia-cuda-cupti-cu12==12.8.90; sys_platform == 'linux'",
    "nvidia-cuda-nvrtc-cu12==12.8.93; sys_platform == 'linux'",
    "nvidia-cuda-runtime-cu12==12.8.90; sys_platform == 'linux'",
    "nvidia-cudnn-cu12==9.10.2.21; sys_platform == 'linux'",
    "nvidia-cufft-cu12==11.3.3.83; sys_platform == 'linux'",
    "nvidia-cufile-cu12==1.13.1.3; sys_platform == 'linux'",
    "nvidia-curand-cu12==10.3.9.90; sys_platform == 'linux'",
    "nvidia-cusolver-cu12==11.7.3.90; sys_platform == 'linux'",
    "nvidia-cusparse-cu12==12.5.8.93; sys_platform == 'linux'",
    "nvidia-cusparselt-cu12==0.7.1; sys_platform == 'linux'",
    "nvidia-nccl-cu12==2.27.5; sys_platform == 'linux'",
    "nvidia-nvjitlink-cu12==12.8.93; sys_platform == 'linux'",
    "nvidia-nvtx-cu12==12.8.90; sys_platform == 'linux'",
]
allow-insecure-host = ["www.paddlepaddle.org.cn"]

[tool.uv.extra-build-dependencies]
clip = ["setuptools==81.0.0"]

[tool.uv.sources]
paddlepaddle = { index = "paddlepaddle" }
paddlepaddle-gpu = { index = "paddlepaddle-gpu" }
clip = { git = "https://github.com/openai/CLIP.git" }

[[tool.uv.index]]
name = "paddlepaddle"
url = "https://www.paddlepaddle.org.cn/packages/stable/cpu/"
explicit = true

[[tool.uv.index]]
name = "paddlepaddle-gpu"
url = "https://www.paddlepaddle.org.cn/packages/stable/cu126/"
explicit = true

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
# 腾讯镜像源，同步自pypi
url = "https://mirrors.tencent.com/pypi/simple"
name = "mirrors_pypi"
default = true

[[tool.uv.index]]
# 腾讯镜像源，同步自pypi
url = "https://mirrors.tencent.com/repository/pypi/tencent_pypi/simple"
name = "tencent_pypi"

[[tool.uv.index]]
url = "https://pypi.org/simple"
name = "pypi"

[dependency-groups]
dev = [
    "memory-profiler>=0.61.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
